+++ src/app/utils/project-detail-modal.util.ts
@@ -26,11 +26,34 @@ export class ProjectDetailModalUtility {
    * @param id project id.
    * @param name project name
    */
-  public openProjectModal(id: number, name: string, returnPage: string): void {
+  public openProjectModal(id: number, name: string, returnPage: string, activeTab?: string): void {
     name = name.split(' ').join('-');
 
-    this.createProjectModal(id, returnPage);
-    this.location.replaceState(`/project/details/${id}-${name}`);
+    if (activeTab) {
+      this.createProjectModal(id, returnPage, activeTab);
+    } else {
+      this.createProjectModal(id, returnPage, 'description');
+    }
+    const newUrl = name ? `/project/details/${id}-${name}` : `/project/details/${id}`;
+    this.location.replaceState(newUrl);
+  }
+
+    /**
+   * Method to update like-count in background of modal
+   * @param id project id.
+   * @param element the object that needs to be updated in background
+   */
+  public subscribeToLikes(id: number, element) {
+    this.modalRef.content.onLike.subscribe(isLiked => {
+      const projectIndexToUpdate = element.findIndex(project => project.id === id);
+      if (isLiked) {
+        element[projectIndexToUpdate].likeCount++;
+        element[projectIndexToUpdate].userHasLikedProject = true;
+      } else {
+        element[projectIndexToUpdate].likeCount--;
+        element[projectIndexToUpdate].userHasLikedProject = false;
+      }
+    });
   }
 
   /**
@@ -38,7 +61,7 @@ export class ProjectDetailModalUtility {
    * @param projectId the id of the project that should be shown.
    * @param activeTab Define the active tab
    */
-  private createProjectModal(projectId: number, returnPage: string, activeTab: string = 'description') {
+  private createProjectModal(projectId: number, returnPage: string, activeTab: string) {
     const initialState = {
       projectId: projectId,
       activeTab: activeTab,